+++ electron/Index.vue
@@ -320,6 +320,8 @@
             try {
                 if(!this.saveLogin) await keyStore.deletePassword('f-list.net', this.settings.account);
 
+                core.siteSession.setCredentials(this.settings.account, this.password);
+
                 const data = <{ticket?: string, error: string, characters: {[key: string]: number}, default_character: number}>
                     (await Axios.post('https://www.f-list.net/json/getApiTicket.php', qs.stringify({
                         account: this.settings.account, password: this.password, no_friends: true, no_bookmarks: true,